Bug - unable to handle login
