+++ backend/src/modules/authentication/config/logging.config.ts
@@ -1,4 +1,6 @@
-const info = (namespace: string, message: string, object?: any) => {
+type loggerFunction = (namespace: string, message: string, object?: any) => void;
+
+const info: loggerFunction = (namespace, message, object) => {
     if (object) {
         console.info(`[${getTimeStamp()}] [INFO] [${namespace}] ${message}`, object);
     } else {
@@ -6,7 +8,7 @@ const info = (namespace: string, message: string, object?: any) => {
     }
 };
 
-const warn = (namespace: string, message: string, object?: any) => {
+const warn: loggerFunction = (namespace, message, object) => {
     if (object) {
         console.warn(`[${getTimeStamp()}] [WARN] [${namespace}] ${message}`, object);
     } else {
@@ -14,7 +16,7 @@ const warn = (namespace: string, message: string, object?: any) => {
     }
 };
 
-const error = (namespace: string, message: string, object?: any) => {
+const error: loggerFunction = (namespace, message, object) => {
     if (object) {
         console.error(`[${getTimeStamp()}] [ERROR] [${namespace}] ${message}`, object);
     } else {
@@ -22,7 +24,7 @@ const error = (namespace: string, message: string, object?: any) => {
     }
 };
 
-const debug = (namespace: string, message: string, object?: any) => {
+const debug: loggerFunction = (namespace, message, object) => {
     if (object) {
         console.debug(`[${getTimeStamp()}] [DEBUG] [${namespace}] ${message}`, object);
     } else {
