+++ backend/src/modules/authentication/interfaces/authRequest.interface.ts
@@ -1,0 +1,9 @@
+interface LoginReq {
+    email: string;
+    password: string;
+}
+
+
+export {
+    LoginReq
+}
