+++ backend/src/modules/authentication/middleware/authorization.ts
@@ -1,16 +1,16 @@
 import jwt from "jsonwebtoken";
+import getErrorMessage from "../../../../errorHandler";
 import dotenv from "dotenv";
 import { resolve } from "path";
 import { verifyJWT } from "../utils/JWT-helpers";
-import getErrorMessage from "../../../../errorHandler";
-import User from '../interfaces/user.interface';
 import { Request, Response, NextFunction } from "express";
 import logging from "../config/logging.config";
-import { error } from "console";
 
 dotenv.config({path: resolve(__dirname, "../../../../.env")});
 
-function authenticateToken( req: Request , res: Response, next : NextFunction ){
+const NAMESPACE = "Authorization";
+
+const authenticateToken = ( req: Request , res: Response, next : NextFunction ) => {
     const authHeader = req.headers['authorization']; // contains "Bearer TOKEN"
     if (authHeader == null) {
         return res.status(401).json({error: "Null token received."});
@@ -21,12 +21,11 @@ function authenticateToken( req: Request , res: Response, next : NextFunction ){
         res.locals.verified = accessDecoded;
 
     } catch (error) {
+        logging.error(NAMESPACE, getErrorMessage(error), error);
         return res.status(401).json({
-            message: "Token validation error:",
-            error: error
+            message: "Token validation error!"
         });
     }
-    
     return next(); 
 };
