+++ backend/src/modules/authentication/routes/auth-routes.ts
@@ -1,14 +1,26 @@
 import express, { Express, Request, Response, Router } from 'express';
-import controller from '../controllers/user';
+import handler from '../handlers/auth';
+import { routerEnclose } from '../utils/routerEnclose';
 import { extractBothJWT, extractRefreshJWT } from '../middleware/extractJWT';
+import { LoginReq } from '../interfaces/authRequest.interface';
 
 const router = Router();
 
-// handle login
+router.post('/register', handler.register);
 
-router.get('/validate', extractBothJWT, controller.validateToken);
-router.get('/refresh', extractRefreshJWT, controller.refreshAccessToken)
-router.post('/login', controller.loginUser);
+router.post(
+    '/login', 
+    routerEnclose(handler.loginUser, ( req: Request, res: Response ) => {
+        const body: LoginReq = req.body; 
+        return {
+            source: "express",
+            payload: body
+        }
+    })
+);
+
+router.get('/validate', extractBothJWT, handler.validateToken);
+router.get('/refresh', extractRefreshJWT, handler.refreshAccessToken)
 
 
 export default router;
